refactor(model-list): move account source catalog loaders
